backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eve-station-trader-backend"
version = "0.1.0"
description = "Backend for EVE Station Trader"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "apscheduler>=3.10.4",
  "fastapi>=0.115.8",
  "httpx>=0.28.1",
  "psycopg[binary]>=3.2.5",
  "pydantic-settings>=2.8.1",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.38",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.18.2",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
  "ruff>=0.14.1",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[[tool.mypy.overrides]]
module = [
  "apscheduler.*",
]
ignore_missing_imports = true
